Test of Equality of Covariances given by Chaipitak and Chongcharoen 2013

Performs 2 and k sample equality of covariance matrix test using Chaipitak and Chongcharoen 2013

Chaipitak2013_test(x, ...)

# S3 method for data.frame
Chaipitak2013_test(x, group, ...)

# S3 method for grouped_df
Chaipitak2013_test(x, ...)

# S3 method for resample
Chaipitak2013_test(x, ...)

# S3 method for list
Chaipitak2013_test(x, ...)

Arguments

x

data as data.frame, grouped_df, resample or matrix object

...

other options passed to functions

group

group or population variable

Value

Test statistic of the hypothesis test

References

Chaipitak, S. and Chongcharoen, S. (2013). A test for testing the equality of two covariance matrices for high-dimensional data. Journal of Applied Sciences, 13(2):270-277.

Examples

Chaipitak2013_test(iris, group = Species)

#> 
#> 	Chaipitak and Chongchareon 2013 Equality of Covariance Test
#> 
#> data:  setosa, versicolor and virginica
#> Chi Squared = 164.41, df = 2, p-value < 2.2e-16
#> alternative hypothesis: true difference in covariances is not equal to 0
#> sample estimates:
#> $`covariance of setosa`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.12424898 0.099216327  0.016355102 0.010330612
#> Sepal.Width    0.09921633 0.143689796  0.011697959 0.009297959
#> Petal.Length   0.01635510 0.011697959  0.030159184 0.006069388
#> Petal.Width    0.01033061 0.009297959  0.006069388 0.011106122
#> 
#> $`covariance of versicolor`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.26643265  0.08518367   0.18289796  0.05577959
#> Sepal.Width    0.08518367  0.09846939   0.08265306  0.04120408
#> Petal.Length   0.18289796  0.08265306   0.22081633  0.07310204
#> Petal.Width    0.05577959  0.04120408   0.07310204  0.03910612
#> 
#> $`covariance of virginica`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.40434286  0.09376327   0.30328980  0.04909388
#> Sepal.Width    0.09376327  0.10400408   0.07137959  0.04762857
#> Petal.Length   0.30328980  0.07137959   0.30458776  0.04882449
#> Petal.Width    0.04909388  0.04762857   0.04882449  0.07543265
#> 
#>