Test of Equality of Covariances given by Schott 2007

Performs 2 and k sample equality of covariance matrix test using Schott 2007

Schott2007_test(x, ...)

Arguments

x
data as data.frame, grouped_df, resample or matrix object

...
other options passed to functions

Value

Test statistic of the hypothesis test

Details

Consider the test for the equality of covariance matrices with $$\text{H}_0: \boldsymbol{\Sigma}_1 = \boldsymbol{\Sigma}_2 = \ldots = \boldsymbol{\Sigma}_k$$ and $$\text{H}_{\text{A}}: \boldsymbol{\Sigma}_1 \neq \boldsymbol{\Sigma}_2 \neq \ldots \neq \boldsymbol{\Sigma}_k,$$ where \(\boldsymbol{\Sigma_i}\) are the population covariance matrix parameters. The sample covarinance matrix estimators, \(\textbf{S}_1, \textbf{S}_2, \ldots, \textbf{S}_i\), are distributed singular Wishart such that \(n_i\textbf{S}_i \sim W_k\left( \boldsymbol{\Sigma}_i, n_i \right)\). Then, $$T_{Sc} = \sum \limits^{k}_{i < j} \frac{ \left( \hat{a}_{2i_{Sc}} + \hat{a}_{2j_{Sc}} - \frac{2}{p}tr \left( S_i S_j \right) \right) ^ 2}{\theta_{Sc}}$$ where, $$\hat{a}_{2i_{Sc}} = \frac{tr \left( \boldsymbol{V}_i^2 \right) - \frac{1}{n_i}tr \left( \boldsymbol{V}_i \right)^2}{ \left( n_i - 1 \right) \left(n_i + 2 \right)p} \xrightarrow{P} \frac{tr \left( \boldsymbol{\Sigma}_i^2 \right) }{p}$$ and $$\hat{a}_{2_{Sc}} = \frac{\sum \limits^k_{i = 1}(n_i -1)\hat{a}_{2i_{Sc}}}{\sum \limits^k_{i = 1}(n_i -1)}.$$ The divisor $$\theta_{Sc} = 4 \hat{a}_{2_{Sc}}^2 \left( \sum \limits_{i < j}^ k \left( \frac{1}{n_i} + \frac{1}{n_j} \right) + (k - 1)(k - 2) \sum \limits_{i = 1}^k n_i^{-2} \right)$$ is a variance term for the numerator.

References

Schott, J. (2007). A test for the equality of covariance matrices when the dimension is large relative to the sample sizes. Computational Statistics & Data Analysis, 51(12):6535-6542.

Examples

Schott2007_test(iris, group = Species)

#> 
#> 	Schott 2007 Equality of Covariance Test
#> 
#> data:  setosa, versicolor and virginica
#> Chi Squared = 61193, df = 1, p-value < 2.2e-16
#> alternative hypothesis: true difference in covariances is not equal to 0
#> sample estimates:
#> $`covariance of setosa`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.12424898 0.099216327  0.016355102 0.010330612
#> Sepal.Width    0.09921633 0.143689796  0.011697959 0.009297959
#> Petal.Length   0.01635510 0.011697959  0.030159184 0.006069388
#> Petal.Width    0.01033061 0.009297959  0.006069388 0.011106122
#> 
#> $`covariance of versicolor`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.26643265  0.08518367   0.18289796  0.05577959
#> Sepal.Width    0.08518367  0.09846939   0.08265306  0.04120408
#> Petal.Length   0.18289796  0.08265306   0.22081633  0.07310204
#> Petal.Width    0.05577959  0.04120408   0.07310204  0.03910612
#> 
#> $`covariance of virginica`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.40434286  0.09376327   0.30328980  0.04909388
#> Sepal.Width    0.09376327  0.10400408   0.07137959  0.04762857
#> Petal.Length   0.30328980  0.07137959   0.30458776  0.04882449
#> Petal.Width    0.04909388  0.04762857   0.04882449  0.07543265
#> 
#>